Get all tests passing (by skipping obsolete ones).
Clean up a tad. Add prospective grammar for Turtles' multi-block sentence syntax.

Remove a newline that was causing a test_lex case to fail. Most likely we'll want to remove all newline tokens that are redundant because of a following dent, but I don't dare risk it until we have the grammar under test.